requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.6
Django==4.2.3
django[argon2]
django-cors-headers==4.1.0
django-environ==0.10.0
django-filter==23.2
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.6
filelock==3.12.2
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.3
MarkupSafe==2.1.3
openpyxl==3.1.2
packaging==23.1
platformdirs==3.8.0
psycopg==3.1.9
psycopg-binary==3.1.9
psycopg2==2.9.6
PyJWT==2.7.0
python-decouple==3.8
python-dateutil>=2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
virtualenv==20.23.1
whitenoise==6.4.0